Change file structure and add new components
