+++ CampusNavigator/CampusNavigator/CampusNavigator/Components/AnnouncementTile.swift
@@ -1,8 +1,36 @@
-//
-//  AnnouncementTile.swift
-//  CampusNavigator
-//
-//  Created by Sasindu 034 on 2025-02-23.
-//
+import SwiftUI
 
-import Foundation
+struct AnnouncementTile: View {
+    let icon: String
+    let title: String
+    let time: String
+    
+    var body: some View {
+        HStack(spacing: 16) {
+            Image(systemName: icon)
+                .font(.title3)
+                .symbolRenderingMode(.multicolor)
+                .frame(width: 40)
+                .foregroundColor(.orange)
+            
+            VStack(alignment: .leading, spacing: 4) {
+                Text(title)
+                    .font(.subheadline.weight(.medium))
+                    .foregroundColor(.orange)
+                
+                Text(time)
+                    .font(.footnote)
+                    .foregroundColor(.secondary)
+            }
+            
+            Spacer()
+        }
+        .padding()
+        .background(Color(.secondarySystemGroupedBackground))
+        .clipShape(RoundedRectangle(cornerRadius: 16, style: .continuous))
+        .overlay(
+            RoundedRectangle(cornerRadius: 16, style: .continuous)
+                .stroke(Color.orange, lineWidth: 1)
+        )
+    }
+}
